requirements.txt
asgiref==3.2.10
Django==3.1
django-phone-field==1.8.1
pytz==2020.1
sqlparse==0.3.1
